Separate smart account & modules generation
