+++ .github/scripts/solidity/types.ts
@@ -1,0 +1,7 @@
+export type ParamType = { name: string; type: string; description: string }
+export type DocContent = {
+  contractName: string
+  eventName?: string
+  functionName?: string
+  contents: string
+}
